requirements.txt
matplotlib==3.6.1
numpy==1.23.4
opencv_contrib_python==4.6.0.66
Pillow==9.4.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
